Do not generate output_file in recursive 'preprocess_file' calls
When preprocessing a file we call the function recursively on all included files.
There, however, we only collect all definitions and throw away all other results of the call.
Here, a big performance impact is done by the generation of the output file, where we essentially copy all file lines and replace definitions in them via regex substitution.

After this commit, we do not generate the output file on recursive calls, which is thrown away by the callee anyways.